Moving main files to correct folder
