--- STM32CubeIDE/Debug/Core/Src/main.su
@@ -1,10 +1,0 @@
-../Core/Src/main.c:71:6:IncCounterToPWM	8	static
-../Core/Src/main.c:81:5:main	144	static
-../Core/Src/main.c:186:6:SystemClock_Config	88	static
-../Core/Src/main.c:233:13:MX_ADC1_Init	24	static
-../Core/Src/main.c:292:13:MX_TIM1_Init	104	static
-../Core/Src/main.c:373:13:MX_TIM2_Init	40	static
-../Core/Src/main.c:418:13:MX_TIM3_Init	40	static
-../Core/Src/main.c:463:13:MX_USART1_UART_Init	8	static
-../Core/Src/main.c:511:13:MX_GPIO_Init	48	static
-../Core/Src/main.c:555:6:Error_Handler	8	static,ignoring_inline_asm
