+++ STM32CubeIDE/Debug/src_STM32Kernel/kernel.su
@@ -1,1 +1,10 @@
-../src_STM32Kernel/kernel.c:11:6:STM32Kernel_Main	8	static
+../src_STM32Kernel/kernel.c:71:6:IncCounterToPWM	8	static
+../src_STM32Kernel/kernel.c:81:5:main	144	static
+../src_STM32Kernel/kernel.c:186:6:SystemClock_Config	88	static
+../src_STM32Kernel/kernel.c:233:13:MX_ADC1_Init	24	static
+../src_STM32Kernel/kernel.c:292:13:MX_TIM1_Init	104	static
+../src_STM32Kernel/kernel.c:373:13:MX_TIM2_Init	40	static
+../src_STM32Kernel/kernel.c:418:13:MX_TIM3_Init	40	static
+../src_STM32Kernel/kernel.c:463:13:MX_USART1_UART_Init	8	static
+../src_STM32Kernel/kernel.c:511:13:MX_GPIO_Init	48	static
+../src_STM32Kernel/kernel.c:555:6:Error_Handler	8	static,ignoring_inline_asm
